% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_shadows.R
\name{cast_shadows}
\alias{cast_shadows}
\title{Cast shadows}
\usage{
cast_shadows(dem, SZA, SAA, dl = 0, sombra = dem)
}
\arguments{
\item{dem}{matrix or SpatRaster. Digital elevation model representing terrain elevation
on a regular grid.}

\item{SZA}{SpatRaster or numeric. Solar zenith angle in degrees.}

\item{SAA}{SpatRaster or numeric. Solar azimuth angle in degrees.}

\item{dl}{numeric. Grid spacing. Not needed if \code{dem} is SpatRaster.}

\item{sombra}{Returned matrix or SpatRaster.}
}
\value{
Returns an object of the same class as the input DEM (matrix or SpatRaster),
with values of 0 for cast-shadowed pixels and 1 for not-shaded pixels.
}
\description{
Calculates cast shadows over a matrix or SpatRaster digital elevation model (DEM)
for a given illumination direction.
}
\details{
\code{cast_shadows()} calls a fortran routine called \code{doshade} that scans the DEM in
lines parallel to the sun direction. It compares the projection of the grid cells on a
plane perpendicular to the sun to determine whether they are exposed to direct solar
illumination or shadowed by neighbor objects, such as mountains. See
\insertCite{corripio_2003;textual}{SatRbedo} for details. The \code{doshade} subroutine
was originally written by Javier Corripio and it was shipped with the \code{insol}
package. This function has been optimized and updated to take advantage of the
infrastructure provided by the \code{terra} package.
}
\examples{
library(terra)
dem <- system.file("extdata/athabasca_dem.tif", package = "SatRbedo")
dem <- terra::rast(dem)
SZA <- 49
SAA <- 165
cs <- cast_shadows(dem, SZA, SAA)
plot(cs, col = grey(0:1))

}
\references{
\insertAllCited{}
}
