% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfactor_BRDF.R
\name{cfactor_BRDF}
\alias{cfactor_BRDF}
\title{c-factor BRDF normalization}
\usage{
cfactor_BRDF(SAA, SZA, VAA, VZA, NSZ, br = 1, hb = 2, band)
}
\arguments{
\item{SAA}{SpatRaster or numeric. Solar azimuth angle in degrees.}

\item{SZA}{SpatRaster or numeric. Solar zenith angle in degrees.}

\item{VAA}{SpatRaster or numeric. Satellite sensor azimuth angle in degrees.}

\item{VZA}{SpatRaster or numeric. Satellite sensor zenith angle in degrees.}

\item{NSZ}{numeric. NBAR solar zenith angle. This prescribed \code{NSZ}
is obtained from the satellite product metadata or calculated using
the algorithm and code provided by \insertCite{li_2019;textual}{SatRbedo}.}

\item{br}{numeric. Crown relative height parameter. Defaults to 1.0.}

\item{hb}{numeric. Shape parameter. Defaults to 2.0.}

\item{band}{character. Spectral band to be processed. There are six bands
that can be considered: "blue (0.43-0.45 um)", "green (0.53-0.59 u)",
"red (0.64-0.67 um)", "NIR" (near-infrared, 0.85-0.88 um),
"SWIR" (shortwave-infrared, 1.57-1.65 um), and "SWIR2 (shortwave-infrared, 2.11-2.29 um)".}
}
\value{
Returns a single numeric value or a SpatRaster with the c-factor
values for a spectral band.
}
\description{
Calculates the c-factor proposed by \insertCite{roy_2016;textual}{SatRbedo}
which can be used to normalize surface reflectance data to nadir BRDF adjusted
reflectance (NBAR), or to convert NBAR observations to off-nadir surface reflectance.
The c-factor is calculated based on a spectral kernel-based BRDF model. The mathematical
formulation of the kernels is presented in
\insertCite{@ @wanner_1995, @lucht_2000 and @carlsen_2020;textual}{SatRbedo}.
}
\examples{
# Using data with numeric format
cfactor_BRDF(SAA = 56.58, SZA = 35.6, VAA = 236.4, VZA = 1.00, NSZ = 34.97, band = "NIR")

# Using SpatRaster data
library(terra)
m1 <- matrix(seq(from = 56.54, to = 56.67, length.out = 9), nrow = 3, ncol = 3)
m2 <- matrix(seq(from = 35.54, to = 35.65, length.out = 9), nrow = 3, ncol = 3)
m3 <- matrix(seq(from = 221.05, to = 259.06, length.out = 9), nrow = 3, ncol = 3)
m4 <- matrix(seq(from = 0.53, to = 1.12, length.out = 9), nrow = 3, ncol = 3)
SAA <- terra::rast(m1)
SZA <- terra::rast(m2)
VAA <- terra::rast(m3)
VZA <- terra::rast(m4)
cfactor_BRDF(SAA, SZA, VAA, VZA, NSZ = 34.97, band = "NIR")

}
\references{
\insertAllCited{}
}
