% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albedo_sat.R
\name{f_BRDF}
\alias{f_BRDF}
\title{Anisotropy correction of reflected radiation over melting snow and glacier ice}
\usage{
f_BRDF(SAA, SZA, VAA, VZA, slope, aspect, method, green, NIR, th)
}
\arguments{
\item{SAA}{SpatRaster or numeric. Solar azimuth angle in degrees.}

\item{SZA}{SpatRaster or numeric. Solar zenith angle in degrees.}

\item{VAA}{SpatRaster or numeric. Satellite sensor azimuth angle in degrees.}

\item{VZA}{SpatRaster or numeric. Satellite sensor zenith angle in degrees.}

\item{slope}{SpatRaster or numeric. Surface slope in degrees.}

\item{aspect}{SpatRaster or numeric. Surface aspect in degrees.}

\item{method}{character. Number of spectral bands considered for the anisotropy
correction. There are two options available: "twobands" (green and NIR) using
the parameterizations of \insertCite{greuell_1999;textual}{SatRbedo} for ice
and \insertCite{koks_2001;textual}{SatRbedo} for snow, and "fivebands"
(blue, red, NIR, SWIR1, and SWIR2) using the methods described in
\insertCite{ren_2021;textual}{SatRbedo}.}

\item{green}{SpatRaster. Green band surface reflectance (0.53-0.59 um).}

\item{NIR}{SpatRaster. Near-infrared band surface reflectance (0.85-0.88 um).}

\item{th}{numeric. NDSII threshold to discriminate between snow and ice.}
}
\value{
Returns a SpatRaster with two layers for \code{method="twobands"} and five layers
for \code{method="fivebands"}. Each layer contains the anisotropic reflection factors (\code{f})
for each spectral band and surface type (snow or ice). All \code{f} values represent the
difference between the satellite albedo retrievals and the surface reflectance observations.
}
\description{
This tool corrects for the anisotropy of reflected radiation by melting snow and ice.
For a given surface type, the correction is carried out using an empirical model
of the Bidirectional Reflectance Distribution Function (BRDF) that depends on
the wavelength bands and the view-solar geometry. The procedure is applicable to visible,
near-infrared (NIR), and shortwave-infrared (SWIR) wavelenghts. More details about
the parameterizations can be found in the references list.
\insertNoCite{greuell_1999,koks_2001,de_ruyter_de_wildt_2002,ren_2021}{SatRbedo}
}
\section{Notes}{

\code{f} is not calculated for \code{method="twobands"} if the green or the NIR surface
reflectance values are greater than one or for pixels for which the SZA relative to the inclined
surface is >65.51째 over snow or >74.2째 over ice. For \code{method="fivebands"}, \code{f} is not
calculated if the SZA relative to the inclined surface is >70.9째 over snow or >57.6째
over ice.
}

\examples{
library(terra)
outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
green <- system.file("extdata/athabasca_2020253_B03_S30.tif", package = "SatRbedo")
nir <- system.file("extdata/athabasca_2020253_B8A_S30.tif", package = "SatRbedo")
dem <- system.file("extdata/athabasca_dem.tif", package = "SatRbedo")
green <- preproc(grd = green, outline = outline)
nir <- preproc(grd = nir, outline = outline)
dem <- preproc(grd = dem, outline = outline)
SAA <- 164.8
SZA <- 48.9
VAA <- 287.1
VZA <- 7.2
slope <- terra::terrain(dem, v = "slope", neighbors = 4, unit = "degrees")
aspect <- terra::terrain(dem, v = "aspect", neighbors = 4, unit = "degrees")

# f coefficients for two bands
th <- snow_or_ice(green, nir)$th
f <- f_BRDF(
  SAA = SAA, SZA = SZA, VAA = VAA, VZA = VZA,
  slope = slope, aspect = aspect,
  method = "twobands", green = green, NIR = nir,
  th = th
)
plot(f[[1]])
plot(f[[2]])

# f coefficients for five bands
th <- snow_or_ice(green, nir)$th
f <- f_BRDF(
  SAA = SAA, SZA = SZA, VAA = VAA, VZA = VZA,
  slope = slope, aspect = aspect,
  method = "fivebands", green = green, NIR = nir,
  th = th
)
plot(f[[1]])
plot(f[[2]])
plot(f[[3]])
plot(f[[4]])
plot(f[[5]])

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=preproc]{preproc()}}, \code{\link[terra:terrain]{terra::terrain()}}, \code{\link[=snow_or_ice]{snow_or_ice()}}, \code{\link[terra:plot]{terra::plot()}}
}
