% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillshading.R
\name{hill_shade}
\alias{hill_shade}
\title{Topographic hillshading}
\usage{
hill_shade(dem, SZA, SAA)
}
\arguments{
\item{dem}{SpatRaster. Digital elevation model representing terrain elevation
on a regular grid.}

\item{SZA}{SpatRaster or numeric. Solar zenith angle in degrees.}

\item{SAA}{SpatRaster or numeric. Solar azimuth angle in degrees.}
}
\value{
Returns a hillshade map in SpatRaster format.
}
\description{
Computes the hypothetical illumination of a surface by setting the position of
the sun and calculating the illumination values for each grid cell of a DEM. The
\code{hill_shade} function is based upon the vectorial algebra algorithms developed
by \insertCite{corripio_2003;textual}{SatRbedo}. This function has been
optimized and updated to take advantage of the infrastructure provided by the \code{terra} package.
}
\examples{
library(terra)
dem <- system.file("extdata/athabasca_dem.tif", package = "SatRbedo")
dem <- terra::rast(dem)
SZA <- 49
SAA <- 165
hs <- hill_shade(dem, SZA, SAA)
plot(hs, col = grey(1:100 / 100))

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=normal_vector]{normal_vector()}}, \code{\link[=sun_vector]{sun_vector()}}
}
