% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_or_ice.R
\name{NDSII_hist}
\alias{NDSII_hist}
\title{NDSII histogram}
\usage{
NDSII_hist(NDSII, th, breaks = 64, stdev = 2)
}
\arguments{
\item{NDSII}{SpatRaster. Normalized Difference Snow Ice Index (NDSII).}

\item{th}{numeric. NDSII threshold to discriminate between snow and ice.}

\item{breaks}{one of:
    \itemize{
      \item a vector giving the breakpoints between histogram cells,
      \item a function to compute the vector of breakpoints,
      \item a single number giving the number of cells for the histogram,
      \item a character string naming an algorithm to compute the
      number of cells (see \sQuote{Details}),
      \item a function to compute the number of cells.
    }
    In the last three cases the number is a suggestion only; as the
    breakpoints will be set to \code{\link{pretty}} values, the number
    is limited to \code{1e6} (with a warning if it was larger).  If
    \code{breaks} is a function, the \code{x} vector is supplied to it
    as the only argument (and the number of breaks is only limited by
    the amount of available memory).
  }

\item{stdev}{numeric. Standard deviation cutoff value for histogram stretching.}
}
\description{
This function creates an NDSII histogram with a vertical line showing the selected threshold
discriminating snow and ice.
}
\examples{
green <- system.file("extdata/athabasca_B03_20200911.tif", package = "SatRbedo")
nir <- system.file("extdata/athabasca_B8A_20200911.tif", package = "SatRbedo")
outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
green <- preproc(grd = green, outline = outline)
nir <- preproc(grd = nir, outline = outline)
res <- snow_or_ice(green, nir)
NDSII_hist(res$NDSII, res$th, breaks = 16, stdev = 3)

}
\seealso{
\code{\link[=preproc]{preproc()}}, \code{\link[=snow_or_ice]{snow_or_ice()}}, \code{\link[graphics:hist]{graphics::hist()}}
}
