% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_vector.R
\name{normal_vector}
\alias{normal_vector}
\title{Unit normal vector of a surface}
\usage{
normal_vector(dem, dlx = 0, dly = dlx, cArea = FALSE)
}
\arguments{
\item{dem}{matrix or SpatRaster. Digital elevation model representing terrain elevation
on a regular grid.}

\item{dlx}{numeric. Spatial resolution along the \code{x} axis in meters.}

\item{dly}{numeric. Spatial resolution along the \code{y} axis in meters.}

\item{cArea}{logical. If \code{TRUE}, returns the surface area of each grid cell. The default
returns the gradient.}
}
\value{
Returns a 3D array with the same number of rows and columns as the DEM and the three
layers in the third dimension corresponding to the (x, y, z) components of the normal vectors
to each of the grid cells. If cArea is \code{TRUE}, the result is a 2D matrix with the
surface area of each grid cell.
}
\description{
Computes a unit vector normal to each grid cell in a digital elevation model.
}
\details{
Mathematically, a vector normal to an inclined surface, which is represented by
a plane enclosed by four data points \insertCite{@see Fig. 1 of @corripio_2003}{SatRbedo},
is defined as the average of the cross products of the vectors along the sides of the grid
cell. By definition, the length of this vector is equal to the surface area of the grid cell.
This function was originally written by Javier Corripio as part of the \code{insol} package.
The function has been optimized and updated to take advantage of the infrastructure provided
by the \code{terra} package.
}
\references{
\insertAllCited{}
}
\keyword{internal}
