% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc}
\alias{preproc}
\title{Satellite data pre-processing}
\usage{
preproc(
  grd,
  drivers = NULL,
  outline = NULL,
  coords = NULL,
  add_offset = 0,
  scale_factor = 1
)
}
\arguments{
\item{grd}{filename (character) of the raster to be processed, in one of GDAL's raster
\href{https://gdal.org/en/stable/drivers/raster/index.html}{drivers}.}

\item{drivers}{character. GDAL drivers to consider}

\item{outline}{filename (character) of the shapefile containing the outline of the area
of interest, in \code{.shp} format.}

\item{coords}{filename (character) of the raster with the coordinate system definition.
Alternatively, a coordinate reference system (CRS) description can
be provided. In this case, the CRS can be described using the PROJ-string
(e.g., "+proj=longlat +datum=WGS84") or the WKT format (e.g., "EPSG:4326").}

\item{add_offset}{band-specific offset added to each grid value. Defaults to zero.}

\item{scale_factor}{band-specific scale factor applied to each grid value. Defaults to one.}
}
\value{
Returns a pre-processed SpatRaster grid.
}
\description{
\code{preproc()} crops an input grid to a specified extent with a polygon, reprojects
the grid to a specified coordinate system, and converts the data from
integer to floating point using an offset and a scale factor.
}
\examples{
# uncorrected grid
f <- system.file("extdata/athabasca_B03_20200911.tif", package = "SatRbedo")
preproc(grd = f)

# crop grid
g <- system.file("extdata/athabasca_B03_20200911.tif", package = "SatRbedo")
v <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
preproc(grd = g, outline = v)

# crop and reproject grid
g <- system.file("extdata/athabasca_B03_20200911.tif", package = "SatRbedo")
v <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
preproc(grd = g, outline = v, coords = "+proj=longlat +datum=WGS84")

preproc(grd = g, outline = v, coords = "EPSG:4326")

# Transform grid values
library(terra)
g <- system.file("extdata/athabasca_B03_20200911.tif", package = "SatRbedo")
orig <- preproc(grd = g)
minmax(orig)
new <- preproc(grd = g, add_offset = 5, scale_factor = 10)
minmax(new)

}
\seealso{
\code{\link[terra:rast]{terra::rast()}}, \code{\link[terra:ext]{terra::ext()}}, and \code{\link[terra:project]{terra::project()}} which this function wraps.
}
