% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_or_ice.R
\name{snow_or_ice}
\alias{snow_or_ice}
\title{Automatic discrimination of snow and ice cover types}
\usage{
snow_or_ice(green, NIR)
}
\arguments{
\item{green}{SpatRaster. Green band surface reflectance (0.53-0.59 um).}

\item{NIR}{SpatRaster. Near-infrared band surface reflectance (0.85-0.88 um).}
}
\value{
Returns a list with two objects. The first component is an NDSII SpatRaster and the
second component provides the optimal threshold to discriminate between snow and ice surfaces.
}
\description{
Calculates the Normalized Difference Snow Ice Index (NDSII)
\insertCite{keshri_2009}{SatRbedo} to discriminate between
snow and ice surfaces. The discrimination is performed using
an automatic threshold selection method based on the Otsu
algorithm \insertCite{otsu_1979}{SatRbedo}.
}
\examples{
green <- system.file("extdata/athabasca_B03_20200911.tif", package = "SatRbedo")
nir <- system.file("extdata/athabasca_B8A_20200911.tif", package = "SatRbedo")
outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
green <- preproc(grd = green, outline = outline)
nir <- preproc(grd = nir, outline = outline)
res <- snow_or_ice(green, nir)

# Plot NDSII
library(terra)
plot(res$NDSII, range = c(-0.5, 0.5))

# Plot a cover type binary map (snow/ice)
library(terra)
plot(res$NDSII > res$th, type = "classes", col = c("#FFFFC8", "#00407F"), levels = c("snow", "ice"))

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=preproc]{preproc()}}, \code{\link[EBImage:otsu]{EBImage::otsu()}}, \code{\link[terra:plot]{terra::plot()}}
}
