% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillshading.R
\name{sun_vector}
\alias{sun_vector}
\title{Solar vector}
\usage{
sun_vector(zenith, azimuth)
}
\arguments{
\item{zenith}{SpatRaster or numeric. Solar zenith angle in degrees.}

\item{azimuth}{SpatRaster or numeric. Solar azimuth angle in degrees.}
}
\description{
Computes a unit vector in the direction of the sun based on the solar azimuth and zenith angles.
}
\details{
Given the spherical coordinates (radius = 1, zenith, and azimuth angles),
the \code{sun_vector()} function gives the (x, y, z) coordinates of a unit normal vector
pointing towards the sun, where the X-axis is positive eastwards, the Y-axis is positive
southwards, and Z along the Earth radius and positive upwards \insertCite{corripio_2003}{SatRbedo}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
