% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo_corr.R
\name{topo_corr}
\alias{topo_corr}
\title{Topographic correction}
\usage{
topo_corr(band, dem, SAA, SZA, method = "tanrotation", IC_min = NULL)
}
\arguments{
\item{band}{SpatRaster. Spectral band to be corrected.}

\item{dem}{SpatRaster. Digital elevation model representing terrain elevation
on a regular grid.}

\item{SAA}{SpatRaster or numeric. Solar azimuth angle in degrees.}

\item{SZA}{SpatRaster or numeric. Solar zenith angle in degrees.}

\item{method}{character. Topographic correction model to be used. Two options are
available: "tanrotation" \insertCite{@implements the rotation model proposed by @tan_2010}{SatRbedo},
and "ccorrection" \insertCite{@C model, @teillet_1982}{SatRbedo}.}

\item{IC_min}{numeric. Minimum incidence angle to mask out regions of low illumination.}
}
\value{
Returns a list with two objets. The first component is a SpatRaster with two layers
providing the illumination condition (IC) and the corrected surface reflectance, and the
second component contains the values of the empirical parameters obtained by each topographic
correction model.
}
\description{
This tool corrects surface reflectance for the effects of topography. The objective
of the topographic correction is to compensate for the differences in solar
illumination to minimize the variability of observed reflectance for similar
targets and obtain the equivalent reflectance values over flat terrain. The empirical
methods provided here are suitable for mountain environments with rugged topography
and non-Lambertian surface properties. More details about these methods can be found
in the references below. \insertNoCite{riano_2003,tan_2013,chen_2023}{SatRbedo}
}
\examples{
library(terra)
outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
nir <- system.file("extdata/athabasca_2020253_B8A_S30.tif", package = "SatRbedo")
dem <- system.file("extdata/athabasca_dem.tif", package = "SatRbedo")
nir <- preproc(grd = nir)
dem <- preproc(grd = dem)
SAA <- 164.8
SZA <- 48.9

# Topographic correction using method="tanrotation"
corr <- topo_corr(band = nir, dem = dem, SAA = SAA, SZA = SZA, method = "tanrotation")
plot(corr$bands[[1]]) # plot IC
plot(corr$bands[[2]]) # plot corrected surface reflectance

# Topographic correction using method="ccorrection"
corr <- topo_corr(band = nir, dem = dem, SAA = SAA, SZA = SZA, method = "ccorrection", IC_min = 0.3)
plot(corr$bands[[2]]) # plot corrected surface reflectance

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=preproc]{preproc()}}, \code{\link[landsat:topocorr]{landsat::topocorr()}}
}
