% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo_corr.R
\name{topo_splot}
\alias{topo_splot}
\title{Reflectance-IC scatterplot}
\usage{
topo_splot(IC, band, maxcell = 1e+05)
}
\arguments{
\item{IC}{SpatRaster. Illumination condition.}

\item{band}{SpatRaster. Spectral band to be processed.}

\item{maxcell}{positive integer. Maximum number of cells to use for the plot}
}
\description{
This function creates a scatterplot of surface reflectance vs. illumination
condition (IC). The plot can be used to explore the dependency of the original
and the topographically-corrected surface reflectance data on IC.
}
\examples{
outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
nir <- system.file("extdata/athabasca_B8A_20200911.tif", package = "SatRbedo")
dem <- system.file("extdata/athabasca_dem.tif", package = "SatRbedo")
nir <- preproc(grd = nir, outline = outline)
dem <- preproc(grd = dem, outline = outline)
SAA <- 164.8
SZA <- 48.9
corr <- topo_corr(band = nir, dem = dem, SAA = SAA, SZA = SZA, method = "tanrotation")

# Scatterplot of IC vs. uncorrected surface reflectance
topo_splot(corr$bands[[1]], nir)

# Scatterplot of IC vs. topographically-corrected surface reflectance
topo_splot(corr$bands[[1]], corr$bands[[2]])

}
\seealso{
\code{\link[=topo_corr]{topo_corr()}}, \code{\link[graphics:abline]{graphics::abline()}}, \code{\link[stats:lm]{stats::lm()}}
}
